#22 - added user settings and the ability to add/remove root users

+++ frontend/src/components/access/AccessService.js
@@ -2,19 +2,29 @@ import axios from 'axios';
 
 export default {
   getUserRoles(projectId) {
-    return axios.get(`/admin/projects/${projectId}/userRoles`)
+    if (projectId) {
+      return axios.get(`/admin/projects/${projectId}/userRoles`)
+        .then(response => response.data);
+    }
+    return axios.get('/root/rootUsers')
       .then(response => response.data);
   },
   saveUserRole(projectId, userDn, roleName) {
-    return axios.put(`/admin/projects/${projectId}/users/${userDn}/roles/${roleName}`, {
-      userDnVal: userDn,
-      projectIdVal: projectId,
-      roleNameVal: roleName,
-    })
-      .then(response => response.data);
+    if (projectId) {
+      return axios.put(`/admin/projects/${projectId}/users/${userDn}/roles/${roleName}`, {
+        userDnVal: userDn,
+        projectIdVal: projectId,
+        roleNameVal: roleName,
+      })
+        .then(response => response.data);
+    }
+    return axios.put(`/root/addRoot/${userDn}`).then(response => response.data);
   },
   deleteUserRole(projectId, userId, roleName) {
-    return axios.delete(`/admin/projects/${projectId}/users/${userId}/roles/${encodeURIComponent(roleName)}`);
+    if (projectId) {
+      return axios.delete(`/admin/projects/${projectId}/users/${userId}/roles/${encodeURIComponent(roleName)}`);
+    }
+    return axios.delete(`/root/deleteRoot/${userId}`).then(response => response.data);
   },
   getOAuthProviders() {
     return axios.get('/app/oAuthProviders')
